replaces VerificationPinReceiver with PinSmsBroadcastReceiver
