+++ library/build.gradle
@@ -47,6 +47,8 @@ dependencies {
 	implementation 'com.android.support:support-v4:26.1.+'
 	implementation files('libs/ccmp-client.jar')
     implementation files('libs/disklrucache-2.0.2.jar')
+	implementation 'com.google.android.gms:play-services-auth:15.0.1'
+
 }
 
 if (!hasProperty('uploadRepositoryUri')) project.ext.uploadRepositoryUri = ''
